+++ app/src/main/java/com/zillotrix/moneytracker/features/budget/data/local/dto/CategoryTotalDTO.kt
@@ -3,5 +3,5 @@ package com.zillotrix.moneytracker.features.budget.data.local.dto
 data class CategoryTotalDTO(
     val categoryId: Int,
     val categoryName: String,
-    val totalAmount: Double
+    val totalAmount: Long
 )
