+++ app/src/main/java/com/zillotrix/moneytracker/features/budget/domain/model/BudgetOverview.kt
@@ -2,6 +2,6 @@ package com.zillotrix.moneytracker.features.budget.domain.model
 
 data class BudgetOverview(
     val yearMonth: Int,
-    val totalBudget: Double,
+    val totalBudget: Long,
     val categories: Map<String, CategoryTotal>
 )
