+++ app/src/main/java/com/zillotrix/moneytracker/features/budget/domain/model/CategoryTotal.kt
@@ -3,6 +3,6 @@ package com.zillotrix.moneytracker.features.budget.domain.model
 data class CategoryTotal(
     val categoryId: Int,
     val categoryName: String,
-    val totalAmount: Double,
+    val totalAmount: Long,
     val percentage: Double
 )
